Enhance reusability of image selector.
