+++ frontend/src/js/_image-selector.js
@@ -1,0 +1,54 @@
+"use strict";
+
+var imageSelector = function (config) {
+  var preview = document.getElementById(config.preview);
+  var selector = document.getElementById(config.selector);
+  var uploadBtn = document.getElementById(config.uploadBtn);
+  var resetBtn = document.getElementById(config.resetBtn);
+  // This field holds original image information.
+  var defaultImage = document.getElementById(config.defaultImage);
+
+  var defaultUrl = preview.src;
+  var clearSelector = function () {
+    selector.value = '';
+    selector.type = '';
+    selector.type = 'file';
+  };
+
+  // On click add/delete image button.
+  uploadBtn.addEventListener('click', function (eve) {
+    var isAddBtn = uploadBtn.getAttribute('for') === config.selector;
+    if (isAddBtn) {
+      return;
+    }
+    uploadBtn.setAttribute('for', config.selector);
+    !!defaultImage && defaultImage.removeAttribute('name');
+    clearSelector();
+    eve.preventDefault();
+  }, false);
+
+  // On click reset image button.
+  !!resetBtn && resetBtn.addEventListener('click', function (eve) {
+    defaultImage.name = config.defaultImage;
+    preview.src = defaultUrl;
+    uploadBtn.removeAttribute('for');
+    clearSelector();
+  }, false);
+
+  // On load new image.
+  selector.addEventListener('change', function (eve) {
+    var imgFiles = Array.prototype.slice.call(eve.target.files).filter(function (f) {
+      return f.type.match('image.*');
+    });
+    imgFiles.slice(0, 1).forEach(function (f) {
+      var reader = new FileReader();
+      reader.onload = function (e) {
+        preview.src = e.target.result;
+        uploadBtn.removeAttribute('for');
+      };
+      reader.readAsDataURL(f);
+    });
+  }, false);
+};
+
+module.exports = imageSelector;
