+++ frontend/src/js/storeUser_store_edit.js
@@ -1,5 +1,6 @@
 "use strict";
 
+var imageSelector = require('./_image-selector');
 var resetSubcategories = function (key) {
   Array.prototype.slice.call(document.querySelectorAll('[data-subcategory-of]')).forEach(function (x) {
     return x.setAttribute('aria-hidden', 'true');
@@ -41,53 +42,13 @@ mapFrame.addEventListener('load', function () {
 });
 
 /* ======================
- *  Image uploader
+ *  Image selector
  * ====================== */
 
-var preview = document.getElementById('storeImage');
-var defaultUrl = preview.src;
-var selector = document.getElementById('storeImageSelector');
-var uploadBtn = document.getElementById('storeImageLabel');
-var resetBtn = document.getElementById('resetImage');
-// This field holds original image information.
-var defaultImage = document.getElementById('defaultImage');
-var clearSelector = function () {
-  selector.value = '';
-  selector.type = '';
-  selector.type = 'file';
-};
-
-// On click add/delete image button.
-uploadBtn.addEventListener('click', function (eve) {
-  var isAddBtn = uploadBtn.getAttribute('for') === 'storeImageSelector';
-  if (isAddBtn) {
-    return;
-  }
-  uploadBtn.setAttribute('for', 'storeImageSelector');
-  !!defaultImage && defaultImage.removeAttribute('name');
-  clearSelector();
-  eve.preventDefault();
-}, false);
-
-// On click reset image button.
-!!resetBtn && resetBtn.addEventListener('click', function (eve) {
-  defaultImage.name = 'defaultImage';
-  preview.src = defaultUrl;
-  uploadBtn.removeAttribute('for');
-  clearSelector();
-}, false);
-
-// On load new image.
-selector.addEventListener('change', function (eve) {
-  var imgFiles = Array.prototype.slice.call(eve.target.files).filter(function (f) {
-    return f.type.match('image.*');
-  });
-  imgFiles.slice(0, 1).forEach(function (f) {
-    var reader = new FileReader();
-    reader.onload = function (e) {
-      preview.src = e.target.result;
-      uploadBtn.removeAttribute('for');
-    };
-    reader.readAsDataURL(f);
-  });
-}, false);
+imageSelector({
+  preview: 'storeImage',
+  selector: 'storeImageSelector',
+  uploadBtn: 'storeImageLabel',
+  resetBtn: 'resetImage',
+  defaultImage: 'defaultImage'
+});
