+++ frontend/src/pug/storeUser_store_edit.pug
@@ -133,33 +133,26 @@ include _layout
               .storeBody_info_body_row-header
                 | 店舗イメージ画像
               .storeBody_info_body_row-body
-                | %{ case imageUrl }
-                | %{ of Just iUrl }
                 .uploaderWrapper
                   input#storeImageSelector.uploaderCore(
                     name="image" type="file" accept="image/*"
                   )
+                  input#defaultImage(
+                    name="defaultImage" type="hidden" value="\#{fromMaybe mempty defaultImage}"
+                  )
+                  | %{ case imageUrl }
+                  | %{ of Just iUrl }
                   label#storeImageLabel.uploaderLabel
                   button.uploaderReset#resetImage(type="button")
                     | Restore initial image.
                   img#storeImage.uploaderImg.storeBody_info_body_row-body-image(
                     src="\#{iUrl}"
                   )
-                  input#defaultImage(
-                    name="defaultImage" type="hidden" value="\#{fromMaybe mempty defaultImage}"
-                  )
-                | %{ of Nothing }
-                .uploaderWrapper
-                  input#storeImageSelector.uploaderCore(
-                    name="image" type="file" accept="image/*"
-                  )
+                  | %{ of Nothing }
                   label#storeImageLabel.uploaderLabel(for="storeImageSelector")
                   img#storeImage.uploaderImg.storeBody_info_body_row-body-image(
                   )
-                  input#defaultImage(
-                    name="defaultImage" type="hidden" value="\#{fromMaybe mempty defaultImage}"
-                  )
-                | %{ endcase }
+                  | %{ endcase }
             .storeSubmit
               button.btn.outerBtn(type="submit")
                 | この内容で店舗情報を更新する
